update academic script；

+++ update_academic.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
 
-rm -rf themes/academic
+if [ ! -d "themes" ];then
+  mkdir themes
+fi
+
+if [ -d "themes/academic" ];then
+  rm -rf themes/academic
+fi
+
 cp -r ../hugo-academic/ themes/academic
